move date selector component outside of route
reduces footprint of date selector component in the route.tsx file
